requirements.txt
validators~=0.22.0
youtube_dl~=2021.12.17
pytube~=15.0.0
Pillow~=10.1.0
pyinstaller